requirements.txt
PySide6>=6.5.0
mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
deep-translator>=1.11.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
